fix: fold runtime event replay tests into fixtures package limit
Merge event-replay coverage into runtime_execution_test.go and shared
helpers so pkg/factorysessionexecution/fixtures stays within the 15-file cap.
